backend/app/requirements.txt
fastapi
uvicorn[standard]
gunicorn
pandas
numpy
scikit-learn
joblib
sentence-transformers
faiss-cpu
python-dotenv
langchain_google_genai
